+++ src/composables/index.js
@@ -1,4 +1,4 @@
-import { getCurrentInstance, shallowReactive, effectScope } from 'vue'
+import { getCurrentInstance, shallowReactive, effectScope, onUnmounted } from 'vue'
 
 export function useRouter () {
   const i = getCurrentInstance()
@@ -33,8 +33,36 @@ export function useRoute () {
 // TODO:
 // export function useLink () {}
 
-// TODO:
-// export function onBeforeRouteUpdate () {}
+export function onBeforeRouteUpdate (guard) {
+  const i = getCurrentInstance()
+  if (process.env.NODE_ENV !== 'production' && !i) {
+    throwNoCurrentInstance('onBeforeRouteUpdate')
+  }
+
+  const router = useRouter()
+
+  let target = i.proxy
+  // find the nearest routerview to know the depth
+  while (target && target.$vnode && target.$vnode.data && target.$vnode.data.routerViewDepth == null) {
+    target = target.$parent
+  }
+
+  const depth = target && target.$vnode && target.$vnode.data ? target.$vnode.data.routerViewDepth : null
+
+  console.log('found depth', depth)
+
+  // TODO: allow multiple guards?
+  i.proxy.$options.beforeRouteUpdate = guard
+
+  const removeGuard = router.beforeEach((to, from, next) => {
+    // TODO: check it's an update
+    return guard(to, from, next)
+  })
+
+  onUnmounted(removeGuard)
+
+  return removeGuard
+}
 
 // TODO:
 // export function onBeforeRouteLeave () {}
