+++ src/entries/cjs.js
@@ -1,0 +1,3 @@
+import VueRouter from '../router'
+
+export default VueRouter
