+++ src/entries/esm.js
@@ -1,0 +1,10 @@
+import VueRouter from '../router'
+
+export const version = '__VERSION__'
+export { isNavigationFailure, NavigationFailureType } from '../util/errors'
+export { START as START_LOCATION } from '../util/route'
+
+// we can't add the other composables here because people could still be using an older version of vue and that would
+// create a compilation error trying to import from vue
+
+export default VueRouter
